chore: とりあえずコンポーネントに分けた
