+++ package.json
@@ -9,6 +9,7 @@
         "build": "tsc && vite build"
     },
     "dependencies": {
+        "@dnd-kit/core": "^6.0.8",
         "@emoji-mart/data": "^1.1.2",
         "@emoji-mart/react": "^1.1.1",
         "@emotion/react": "^11.10.5",
