+++ src/components/Menu.tsx
@@ -26,30 +26,14 @@ import { ConcurrentLogo } from './ConcurrentLogo'
 import buildTime from '~build/time'
 // @ts-expect-error vite dynamic import
 import { branch } from '~build/info'
+import { StreamList } from './StreamList'
 
 export interface MenuProps {
     streams: string[]
 }
 
 export function Menu(props: MenuProps): JSX.Element {
-    const appData = useContext(ApplicationContext)
-    const [watchStreams, setWatchStreams] = useState<Stream[]>([])
-
     const theme = useTheme<ConcurrentTheme>()
-
-    useEffect(() => {
-        ;(async () => {
-            setWatchStreams(
-                (
-                    await Promise.all(
-                        props.streams.map(
-                            async (id) => await appData.streamDict?.get(id)
-                        )
-                    )
-                ).filter((e) => e) as Stream[]
-            )
-        })()
-    }, [props.streams])
 
     return (
         <Box sx={{ gap: '15px', height: '100%' }}>
@@ -191,53 +175,7 @@ export function Menu(props: MenuProps): JSX.Element {
                     </List>
                 </Box>
                 <Divider />
-                <Box
-                    sx={{
-                        display: 'flex',
-                        flexDirection: 'column',
-                        gap: '5px',
-                        overflowY: 'auto',
-                        overflowX: 'hidden'
-                    }}
-                >
-                    <List
-                        dense
-                        sx={{
-                            width: '100%',
-                            maxWidth: 360,
-                            display: 'flex',
-                            flexDirection: 'column'
-                        }}
-                    >
-                        {watchStreams.map((stream) => {
-                            const labelId = `checkbox-list-secondary-label-${stream.id}`
-                            return (
-                                <ListItem key={stream.id} disablePadding>
-                                    <ListItemButton
-                                        component={Link}
-                                        to={`/#${stream.id}`}
-                                        sx={{ gap: 1 }}
-                                    >
-                                        <PercentIcon
-                                            sx={{
-                                                color: 'background.contrastText'
-                                            }}
-                                        />
-                                        <ListItemText
-                                            id={labelId}
-                                            primary={
-                                                stream.meta
-                                                    ? JSON.parse(stream.meta)
-                                                          .name
-                                                    : 'backrooms'
-                                            }
-                                        />
-                                    </ListItemButton>
-                                </ListItem>
-                            )
-                        })}
-                    </List>
-                </Box>
+                <StreamList streams={props.streams} />
             </Box>
         </Box>
     )
